change: move formatDate to tools.ts
