+++ backend/src/chat/tools.ts
@@ -1,0 +1,14 @@
+/* eslint-disable */
+
+export function formatDate(date: Date): string {
+  const options: Intl.DateTimeFormatOptions = {
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+    hour: '2-digit',
+    minute: '2-digit',
+    hour12: false,
+    timeZone: 'Asia/Tokyo',
+  };
+  return date.toLocaleString('ja-JP', options);
+}
